+++ web-components/src/off-webcomponents.ts
@@ -2,4 +2,8 @@ export { setLocale, getLocale } from "./localization"
 export { OffWebcomponentsConfiguration as OffWebcomponentConfiguration } from "./components/shared/off-webcomponents-configuration"
 export { RobotoffQuestion } from "./components/robotoff-question/robotoff-question"
 export { RobotoffNutrients } from "./components/robotoff-nutrients/robotoff-nutrients"
+
 export { KnowledgePanelsComponent } from "./components/knowledge-panels/knowledge-panels"
+
+export { DonationBanner } from "./components/donation-banner"
+